requirements.txt
boto3
beautifulsoup4
requests

pymongo
python-dotenv
fastapi
uvicorn